+++ src/hydracloud/cloud/network/packet/impl/normal/DisconnectPacket.php
@@ -18,11 +18,7 @@ use hydracloud\cloud\util\terminal\TerminalUtils;
 
 final class DisconnectPacket extends CloudPacket {
 
-    public function __construct(private ?DisconnectReason $disconnectReason = null {
-        get {
-            return $this->disconnectReason;
-        }
-    }) {}
+    public function __construct(private ?DisconnectReason $disconnectReason = null) {}
 
     public function encodePayload(PacketData $packetData): void {
         $packetData->writeDisconnectReason($this->disconnectReason);
@@ -35,16 +31,14 @@ final class DisconnectPacket extends CloudPacket {
     public function handle(ServerClient $client): void {
         if (($server = $client->getServer()) !== null) {
             if ($server->getServerStatus() === ServerStatus::OFFLINE()) {
-                if (isset(CloudServerManager::getInstance()->getAll()[$server->getName()])) {
-                    CloudServerManager::getInstance()->remove($server);
-                }
+                if (isset(CloudServerManager::getInstance()->getAll()[$server->getName()])) CloudServerManager::getInstance()->remove($server);
                 return;
             }
 
             $server->setServerStatus(ServerStatus::OFFLINE());
-            new ServerDisconnectEvent($server)->call();
+            (new ServerDisconnectEvent($server))->call();
             if (CrashChecker::checkCrashed($server, $crashData)) {
-                new ServerCrashEvent($server, $crashData)->call();
+                (new ServerCrashEvent($server, $crashData))->call();
                 CloudLogger::get()->info("The server §b" . $server->getName() . " §ccrashed§r, writing crash file...");
                 CloudServerManager::getInstance()->printServerStackTrace($server->getName(), $crashData);
                 CrashChecker::writeCrashFile($server, $crashData);
@@ -52,17 +46,16 @@ final class DisconnectPacket extends CloudPacket {
                 CloudLogger::get()->success("The server §b" . $server->getName() . " §rhas §cdisconnected §rfrom the cloud.");
             }
 
-            if ($server->getCloudServerData()->processId !== 0) {
-                TerminalUtils::kill($server->getCloudServerData()->processId);
-            }
+            if ($server->getCloudServerData()->getProcessId() !== 0) TerminalUtils::kill($server->getCloudServerData()->getProcessId());
 
             ServerClientCache::getInstance()->remove($server);
             CloudServerManager::getInstance()->remove($server);
-
-            if (!$server->getTemplate()->getSettings()->static) {
-                FileUtils::removeDirectory($server->getPath());
-            }
+            if (!$server->getTemplate()->getSettings()->isStatic()) FileUtils::removeDirectory($server->getPath());
         }
+    }
+
+    public function getDisconnectReason(): ?DisconnectReason {
+        return $this->disconnectReason;
     }
 
     public static function create(DisconnectReason $disconnectReason): self {
